Complexity is Key
I got it to work perfectly! The output is a beautiful Peirce Quincuncial
projection. My main problem before was that this projection requires a
complex function called cn(u,k), and every single definition on the
internet was exceedingly vague. Luckily, I found a package online that
could calculate cn for me, which meant that I didn't need to do any
crazy integration or calculus whatsoever.