Yet More Maps
I created a bunch of new maps! I added the Mollweide projection to the
program, and a Winkel Tripel button that does not yet work. I have
Winkel working in the vectorized version, but I am still working on the
math in the main program. Also, check out the "Shifted" axis preset (now
my favorite projection when combined with "Shifted Quincuncial), as well
as the "Circles", "Stars", and "Political" themes.