I made a map projection!
I invented a new map projection! I have given it the working title
"Rectus Aequilibrium," because it is square, it is a compromise
projection, Latin is cool, and I promised myself a long time ago that if
I ever made a map projection, I would not name it after myself. Feel
free to try it, but I think I will be trying a different algorithm,
since I think this one distorts shape too much to justify the small
amount of size distortion it reduces.
In other news, I implemented a Van der Grinten projection and spruced up
my naming conventions a bit.