Robinson Projection, but Arthur sucks
I made a robinson projection, but if you look extremely closely right
around 45 degrees north, everything is extremely distorted at small
scales. This is because of Runge's phenomenon. Honestly, linear
interpolation would be better than polynomial for this. I don't know
what Arthur Robinson was thinking using Aitken for this. Give me another
commit or so to figure this out.