Threads are evil
I implemented some smarter threading so that error messages block the
main window and the map isn't calculated three times when you choose an
aspect. There was apparently also a critical problem that caused the
vector designer to crash one out of every however many thousand times,
which I fixed with a call to Platform.runLater.
Oh, and I did equidistant conic projections! That probably should have
been the title. I've got this super robust method of dealing with all
the different shapes it can have, which looks utterly hideous but works
amazingly.
In addition to that, I composed my "Map wishlist", which does not at all
describe what it is, but I like the word "wishlist" so whatever. I just
added all the maps I want to have eventually with TODO tags on them.
Once I have all of those, I will consider this part of the code done and
move on to refining all of my analysis tools and generating HTML.