Has this one really not been invented yet?
It's a pseudoazimuthal projection based on stereographic. I call it
"pseudostereographic". It's the next logical step after Aitoff and
Hammer. It's not that it's a particularly good map projection -
obviously scaling the longitudes ruins the conformal properties, and it
comes out as a little more progressive than Plate Caree, just barely
more authagraphic than Lee. It's so obvious, though, that surely someone
must have thought of this already.