It's the final conic! DODODO DOOOO DANANA NA NAAAAAAAA
I implemented the final stages of the conformal conic. It is now alway
centered and always fills the image. I also found and fixed a bug where
the user could change the parameters as the map was being calculated. As
predicted, the code for Lambert conic now looks just as ugly as ED Conic
and Albers, but that's okay, because I never have to touch them again.
Woo!