I'm Ant-Man!
I learned (sort of) how to use Ant and did a bunch of fun build stuff.
Now I have a script to update the .jar files for me and a bunch of real
executables, with installers and icons and everything! I also finally
just got rid of the wormhole thing; it's just completely different from
the rest of this repository and does not belong here. It's fine; I've
got it in the history.