One Retroazimuth to rule them all
I combined the front and back hemispheres of the Hammer Retroazimuthal
to make one crazy looking map projection. It is still technically
correct, even though I rotated half of it 180 degrees to make it fit. As
classy as it was to have one map projection in two pieces, I think it
just works a lot better this way.