Two azimuths are better than one
I implemented the vector version of 2-point equisdistant, and while I
was at it, I changed up the vector version of obliquify. It is a bit
harder to read now, but somewhat faster, and no longer requires the
Vector utility class. I was going to do the raster version next, but
I've just had the most brilliant idea for refactoring the Projection
enum (maybe. It might be complete antidisestablismentarianism), and I
really want to see if Java will let me do it.