Better than Atlantis
I added a new aspect, roughly based on the "Atlantis Projection", which is an oblicque Mollweide projection, except that I recently discovered that that one is transverse, not completely oblique. What is it with this aversion the world has to oblique projections? Anyway, I came up with a slightly better one in a similar location.