Meridians and parallels intertwined like wild vines
I added a graticule to raster maps! This took quite a while to complete, and it still has major issues. Namely, those annoying horizontal lines where it crosses from one side of the map to the other. I've contemplated several methods for dealing with that, and I'm starting to like an intelligent system that knows where the interruptions are and breaks lines that cross it. I don't have the details worked out. This also took a while, so I might take a break from that and finally get around to doing the inverse projection for Cahill-Keyes.