I knew Keyes was good, but... wow.
I can't express in text how impressed I am by Gene Keyes. Perhaps that's
unnecessary. Perhaps it's just the interruption that's making it look so
good. But either way, wow! It's such a good projection. I didn't want to
pass judgement until I had gotten a complete MapAnalyzer color plot, in
case my eyes deceived me. The results are in, and this is a magnificent
map, second only to Lemons (I found out the real word for those is
"Gores", but I'm going to keep using "Lemons"). Cahill-Keyes is complete
and ready for production!