Wow, that was bad.
I found a bug where maps that had no aspect would still use the aspect of the previous map, and to fix it, the user would have to go back to a map with an aspect, set it to standard, and then go back and reload the aspectless one. That has now been corrected.