I thought I'd gotten this one
I'm actually a big fan of Equal Earth and want it to accomplish its sworn mission to destroy the Gall-Peters projection. I can't believe I didn't have it in my program before. The polynomial proved to be surprisingly complex to implement, particularly as far as getting the map's width and height, but I came up with a nice, elegant solution.