There's no way to change this threshold
All I wanted was to hide the category labels for every other bar in my bar chart. It does that automatically when it gets resized, and it had been annoyingly irregularly hiding some of my labels already. Is there an option for the threshold at which that happens, or to change the manner in which it chooses which ones to hide? No. Is there a list of Labels for the ticks that I can access and manually edit? No. I'll just set every other x value to the empy string, then. JavaFX takes it upon itself to sum the y values of all bars with matching x values and combine them into one bar.

I finally had to just make the chart smaller to trigger the automatic reformatting. Lord Ruler, I hate this.