Elastic Earth projections
I implemented the Elastic projections!  this is just a first pass, so of course there are issues to be worked out.  the oceanic one has a few particular longitudes that it won't project, and the gradient calculation seems to be rong near the section edges.  I also haven't synckd up the section gradients at seams.  overall, tho, I'm surprised how good I got the gradient calculation, and the interpolation in general, on the first try.  I also haven't done any of the inverse projection yet, but I intend to do a better job with these ones than I did with Danseiji's inverse projections.

I also did some light refactoring; Meshed -> Danseiji and setParameters -> initialize, and +docstrings, so that accounts for most of the modified files.