smooth dateline
I made the projection smooth at the antimeridian, because it was bothering me.  it looks great now.  well, I did also notice that a bunch of points near the poles are going to NaN in the vector mapdesiner now, but that doesn't seem to be related to this change, so I'll deal with it later.