don't modify parameters
the problem was that I was modifying the initial gess rather than copying it, and that gess was getting passd to multiple sections and then ultimately returnd as the anser.  now it clones it when it starts the iteratinos and then copies it agen when returning it, so the problem went away.

I also made the cropAtPi checkbox work, which was onestly the thing I was doing in the first place (I had no idea this was the root of the problem).  and I removed the projected border because I realized I didn't actually need it to implement cropAtPi