everyone loves nested optimization
I added a backstepping line search to each iteration of my Newton-Raphson solver.  it actually gets rid of most of those rifts in Elastic Earth III where the folding was causing Newton-Raphson to have difficulty.  but there are still a few spots that are struggling.  also, it turns out a few points are still not getting counted in any polygon.