Levenberg–Marquardt
the simple Gauss–Newton with Armije backstep was insufficiently robust, so I implemented LM.  it seems to work better, but it's still not good enuff.  the step size doesn't turn as much as I would expect when λ goes up.  I think next it's time to implement the geodesic acceleration correction term.