write input maps directly
I refactord the map generation code.  it now updates the SVG files automaticly rather than waiting for the user to copy and paste the output into them.  I did this because I can't imagine doing that for all of them and can't beleve I used to think that was an acceptable waste of my own time.  naturally this caused some updates to the data and improvements to the code to get pushd to the SVGs.  most notably, the natural earth dataset reflects that the aral sea is gone.  I didn't update Supermap.svg because it turns out that one required a dataset that I can't find anywhere online.  I have that file still so I could update it, but I would really rather my work be reproducible, and am mad at myself for not documenting that file at all when I found it.  I'd like to get it working with the natural earth populated places shapefile, but we'll see.