applied circle spacing
I set it up to operate the circle spacing algorithm on real SVG files.  I've done a test on this equirectangular map, tho obviusly I intend to actually use this on the elastic projections.  it's annoyingly slow rite now so I need to find a way to speed it up.  I think supplying gradients will help a lot with that.