refactor SVGMap.java
so, I actually did this a while ago and then was afraid to test it.  so... I don't remember exactly what I changed.  I *think* in principle this is supposed to be the exact same behavior as before but more readable and slitely more efficient.  I think the main improvement is that currentFormatString is now only modified in the inherited functions and not hidden in the private functions.