revise Supermap.svg styling
not sure why I'm putting so much work into an input may I may never use but whatever.  the main thing to note here is that I've done away with the use of label rank to determine label size, because as it turns out that's not at all what it's supposed to do.