simplify gradient estimation in Elastik.java
I set the gradient calculations at edges of the defined spline area to always use first-order differentiation rather than twoth-order.  this was originally supposed to help fix my intersecting spline surfaces.  it didn't but I'm keeping it because it does make the code much simpler and the difference seems to be imperceptible.