implement asymmetric bounding boxen
I replaced width and height attributes of map projections with a boundingbox attribute that contains more information – specificly translation.  previusly all map projections needed to have bounding boxen centerd on the origin, and this was quite the headake for the math in some of them.  this change simplifies many of those, and also opens the door for further simplificatiosn when I (as I intend to do) implement a class that automaticly calculates the boundary of a lenticular map projection rather than requiring that I specify it.  more broadly, this opens the door for even more specifications of map shape, which I intend to implement and pursue for the purpose of generating envelope elements as a matter of course for SVG maps.