refactord polyhedral code to use real objects
the old way I was doing it with arrays with comments indicating what each element of each array represented was terrible, and I'm mad at past me about it.  this is way better, and will expand better when I start doing polygon map boundaries.  it took me so long to figure this out because while debugging a relatively simple issue I experimentally swapd out the Van Leeuwen configuration and then swapd it back incorrectly and then was confused for several ours.