rewrite SVGMap class
I refactord it to make the SVGMap class much more logical and expandable.  in particular, rather than a list of strings and a corresponding list of paths that interweave with the strings, and some of the paths are actually circles represented with made-up path commands, it's now a single heterogeneus list of strings, paths, and points all represented differently.  this turnd out to be somewhat harder than I expected because in the process I realized I didn't really understand the flow of information in the existing code (because it was really convoluted) and ended up refactoring when all of the coordinate system conversions and stuff happen.  now the SVGMap class is more or less encapsulated; it has all of the information needed to represent a map in the SVG format, and can just be saved at any time (previusly the save method required the passing of another SVGMap instance because only part of the object was getting passed around most of the time...).  when you do the projection, you create a new projected SVG map, and then save it.  it's extremely simple.  if only making it so hadn't