fix broken ISO 2166 codes
the NaturalEarth database marks the Spratly islands with the 3-letter code "PGA" and the 2-letter code "PG" for some reason.  it seems I previusly mistook it for Papua New Guinea (since it's "PG" in 2-letter codes) and deleted it since Papua New Guinea is PNG, but that causes the political map generation to fail.  so I put PGA back and now its 2-letter code is "SP".