(rectangular) Backgrounds!
I made it so that background rectangles in the input SVGs fit themselves to the bounding box of the new projection in the output SVGs.  this may seem a strange amount of work to go thru seeing as the result is effectively the same as what I got before using <rect x="0" y="0" width="100%" height="100%" />.  but of course this is bilding toward a more general background shape implementation.  this is a necessary groundwork for that advance – now rectangles are parsed when the SVG is loaded and respond to the projection when the map is saved.  the input SVGs also have background rectangles that more naturally match the input.  now all that's left to do is implement those shapes!