Form-fitting maps
I implemented shapes!  now when you input a map with a rectangle over the full equirectangular domain, rather than merely fitting it to the output map's dimensions, it conforms to its shape!  it looks wonderful.  I can't beleve I never saw fit to implement this before.  and it will especially help with the Elastic projections.  I ended up implementing this new Path class because there ended up being a lot of Path manipulation in my shape implementation, and things like "translate this path dx dy" and "invert this path in y" turn out to be useful in a lot of parts of my code.

unfortunately this doesn't quite work rite for the octohedral projections just because their shapes are pretty complicated and I pmelemnted them in a way I know to be imperfect.  my pland solution is to break the faceProject() method out into a separate map projection for each projection formula, so I can then use that projection's drawLoxodrome() method to define the shape of the full octohedral one.  (btw did you realize how powerful drawLoxodrome() is?  I should be using this for everything like damn).