don't duplicate work
I set a flag to prevent multiple maps from being calculated at once.  idk how important this will be since I found out why it's currently so slow and intend to fix that, but just in case someone has a slow computer or one of the projections is really intensive, this will prevent the map from jerking around annoyingly as the sliders are dragd.