I'll have you arrayned for this
I replaced all the linkedlists of path commands with arraylists because it was actually making the projection exteremely slow.  I could alternatively have reritten all of my for-loops over paths to be for-each instead of index-based, which would have solved the same problem, but I think arraylists are just better in general for this.  it's not like I'm short on memory here.  there's only one instance of me calling .remove(), and it's right after some funky indexing stuff, so a linkedlist wouldn't really help there anyway.