error messages that you can read
I discoverd that the error pane automaticly cuts off the last few words of your error message if it can fit *almost* the whole message on one line.  god, javafx is stupid.  I hacked at it until it wraps consistently so that *all* – not just *almost all* – of the words I ask it to show the user are actually shown to the user.