simplify SVGs
I tried to make my SVGs more robust in terms of only using features supported on a wide variety of platforms.  in particular, it turns out nested CSS blocks may not be a core CSS thing... they're a Sass thing?  but chrome and firefox both seem to interpret it correctly automaticly.  in any case, I got rid of that and told the user to do their circle showing manually.  I also deleted all these namespaces and metadata tags because it seems like intellij is having trubble resolving their sources and idk what's the point of including them in the first place.  all of this information is either already included or is better given in supplementary files.  I also added more background rectangles.