fix hammer retroazimuthal
the hammer retroazimuthal was rong as it turns out; the back hemisphere was mirrored in x from what it was supposed to be.  I set up a more robust implementation where the front and back hemispheres are now separae projections that you can use, in addition to the full one with the rotate back hemisphere.  part of this was creating a tility function in Shape to combine th eoutlines of the front and back hemispheres, which incidentally turnd out to be super helpful for Waterman's shape definition as well.