improve graticule at poles
my real goal here was to add the outline of the full map to the gradicule, but I realized that would be hard because the Awt package I'm using only supports line segments; not arcs or anything.  so maybe another time, but for now have a slite improvement where the meridians are thinned out rite near the pole.