FINAL VERSION OF ELASTIC PROJECTIONS
I made some final improvements to the map data files and accordingly to the inverse code.  in particular there are nans in the inverse raster now, and an extra check when returning an inverse solution from a local minimum.  it turns out there's kind of a concerning number of local minima in Elastic III, which I think means there's a concerning amount of self-intersection in the spline surface.  it makes me wonder if linear interpolation isn't just better for this one.  but I really don't want to implement two different types of interpolation, so for now this is it.