remembering how to rebuild
I tried to bild the exe files for the next release and found that it didn't work.  I got confused for a really long time.  then I figured it out.  it turns out Java 17 is kind of completely different from Java 8.  I kind of wish I had stayd on Java 8, but I don't want to go back now.  also for some reason it has to be msi now not exe.  I deleted the eclipse files as well because it's outdated and has a bunch of java 8 stuff.