back to square 8
turns out the reason it wasn't working was that I was using newer Java!  I'm a bit confused because I *thaut* I had used Java 17 or something to build this in the past, but I gess for whatever reason I've always used 8 to make the JAR and EXE files.  whatever, what matters is that it *finally* works!