he, she, me, indicatrices.
I dubble-checked the terminology, and actually an indicatrix is a single circle and the plural is indicatrices.  I think I had this exact same revelation in the past, but the end result is that I was rite before.  thankfully, I did also find out that there's a preexisting non-terrible alternative, which is to simply call them Tissot's ellipses.  which is good because now I don't have to come up with my own (I was leaning toward "Tissot's indicators").  I went ahead and changed the terminology to "ellipse" everywhere in the SVGs.