begone, line!
I set the political svgs to remove the divide thru russia at the antimeridian.  I'm a bit surprised I didn't try this sooner since it's just kind of an ugly erroneus line in most of my map projections.  unfortunately, the solution I found doesn't work for the physical maps.  that's a bit of a bummer for the tissot ellipse maps specificly.  however, it was most notable (and most potentially confusing) in the blank political templates, so I'm very glad that I fixd that.