clip maps to background
this is so good.  I must have thaut of this at some point in the past, but it wasn't much of an option until I added backgrounds.  the geographic content now clips to the background, so if you have a map with a lot of interruptions and there are lines that didn't get cut properly or shapes enclosing singularities such that there's gunk extending off the map, it will be invisible.  the clipping path hides it.  it looks incredible instead of always garbage!  it doesn't fix all such issues of course – there are often excess lines inside the geographic region that won't get removed by clipping.  but it gets like a lot of them.  I can't get over how much better these maps look out of the box now than they did  a year ago.