support maps with units
I realized that the supermap has been incompatible with the mapdesignervector for some time now, ever since I set the width to 150cm, because the SVG parser didn't know how to account for units in the width and heit.  I made a simple Quantity class to keep track of the units as well as the value and now it works fine.  now, I did realize that more than just the width and heit can have units.  in principle I need to check every number I parse from the SVG for units. but I don't want to do that rite now.  I think the width and the height are the most likely numbers to have units attached, so I'm calling this fine for now.