remove south africa's provinces
I was at first operating on the principle of the more resolution the better, but then I realized that South Africa is really not a big country.  23rd in population, 24th in area.  I don't know why it has provincial resolution in this dataset and argentina, mexico, algeria, DRC, saudi arabia, kazakhstan, and india don't.  but the fact that it does is slitely insulting.  I'm removing it from the provinces map because I can't imagine a situation where it would make sense to draw these lines but not the others for which I don't have data.