fix issue with unicode in SVG
apparently map projections with unicode in their names create invalid SVG when using the executable (even tho it works fine when I run the source code...).  something about the encoding I gess?  I don't really understand, but it's better to escape the map projection names anyway just to be consistent with the templates that are already escaped.  I did that and now it seems fine.