fix Waterman projection bug
when I redefined the Waterman projection coordinates (I don't even remember when) to use four-times-larger units, I apparently missed *one* constant, and it was increasing the slope of this line between these two pieces of the Waterman face that was causing the projections at that interface to not line up.  goodness, what an asinine bug.  good on me for finding it in less than an our of its reporting, tho!