I hate Adobe Illustrator
apparently in addition to dropping the delimiter before negative numbers, because you can infer from the hyphen that it's the start of a new number, they also drop it before numbers that start with a decimal point iff the previus number has a decimal point in it.  this is awful.  why did they do this.  it can't possibly have been worth it.  whatever, I account for it now.