account for various advanced SVG <path> features
this adds support for the S and T Bezier shortcuts and commands with multiple sets of arguments, and also has better error handling for when the number of arguments is not what we expect.  I also refactord it a bit to be much cleaner in general, because wow was this poorly written.