Liquid Earth projection
I implemented Robert Sargent's new Liquid Earth projection.  it works sort of, but there are some issues I still need to work out.  "Extend map area" is effectively always turned on, there are some bizzare artifacts that I don't think are right, and it's really slow because the caching is messed up.  but this is an excellent start.