add creator's name to projection info
now when you select a map projection in the "More..." menu, you'll see the name of the person who first introduced it in addition to its geometry, special property, and description.  I did this because now that I no longer have an "Invented by Justin" category, I want to make it clearer which are invented by others and which are my own, just as a conflict of interest disclosure thing.