handle a wider range of Lagrange projection parameters
I allowed the user to set the longitude scale of the Lagrange projection as low as 0.01 and as high as 1.5.  in order for this to work at all I needed to make the shape calculation much more complex.  now it will calculate the actual envelope, which can become infinitely large, and crop that to a square that scales to keep Africa at ruffly the same scale.  this may now be one of my most complicated projection shape calculations, which I think is a little funny given that I only decided to generalize this projection on a whim, and it really doesn't seem all that complicated.